+++ Product/MiscConstants.cpp
@@ -7,4 +7,4 @@ const char* FAN_SPEED_NAMES[] = { "Low", "Medium", "High" };
 
 const int fanDutyCycles[] = { 0, 50, 100 };
 
-const unsigned int expectedFanRPM[] = { 7479, 16112, 22271 };
+const uint16_t expectedFanRPM[] = { 7479, 16112, 22271 };
