+++ Product/MySerial.h
@@ -1,3 +1,4 @@
+// This interface gives access to the serial port
 #pragma once
 #include "WString.h"
 
@@ -7,4 +8,4 @@ bool serialActive();
 void serialPrintf(const char* __fmt, ...);
 void serialPrint(const __FlashStringHelper* string);
 void serialPrintln(const __FlashStringHelper* string);
-char* renderLongLong(long long num);
+char* renderLongLong(int64_t num);
